Split settings screen into a modern hub with category sub-screens
The monolithic SettingsScreen scrolled through every option in a single
LazyColumn. Replace it with a hub that lists each settings category as a
card with an icon, label and short description, navigating to a focused
sub-screen for each group:

- Relay info (host, port, name, contact, description, icon)
- Access control (auth toggle + pubkey/kind allowlists)
- Network (boot, Pokey, proxy, Tor)
- Relay aggregator
- Retention (auto-delete + protected pubkeys/kinds)
- Backup
- Web clients

Each sub-screen owns its own state and Apply button so changes are
scoped to that category. The hub keeps a global Reset-to-defaults action
behind a confirmation dialog.

The bottom bar shows a back action on sub-screens, and the top bar shows
the screen's title.
