+++ app/src/main/res/values/strings.xml
@@ -141,4 +141,21 @@
     <string name="relay_aggregator_last_refresh">Last refresh %1$s ago</string>
     <string name="relay_aggregator_never_refreshed">No refresh yet</string>
     <string name="relay_aggregator_signer_login">Log in with external signer</string>
+    <string name="settings_category_relay_info">Relay info</string>
+    <string name="settings_category_relay_info_description">Host, port, name, contact and description</string>
+    <string name="settings_category_access_control">Access control</string>
+    <string name="settings_category_access_control_description">Authentication and pubkey/kind allowlists</string>
+    <string name="settings_category_network">Network</string>
+    <string name="settings_category_network_description">Boot, Pokey, proxy and Tor</string>
+    <string name="settings_category_aggregator">Relay aggregator</string>
+    <string name="settings_category_aggregator_description">Mirror events from upstream relays</string>
+    <string name="settings_category_retention">Retention</string>
+    <string name="settings_category_retention_description">Auto-delete and protected pubkeys/kinds</string>
+    <string name="settings_category_backup">Backup</string>
+    <string name="settings_category_backup_description">Automatic database backups</string>
+    <string name="settings_category_web_clients">Web clients</string>
+    <string name="settings_category_web_clients_description">Host static web apps on the relay</string>
+    <string name="reset_all_settings">Reset all settings</string>
+    <string name="reset_all_settings_warning">All settings will be restored to their default values. This action cannot be undone.</string>
+    <string name="settings_reset_complete">Settings reset to defaults</string>
 </resources>
